# docs/core/openapi.yaml
openapi: 3.1.0
info:
  title: ecommerce-aigc-video API
  version: 2.0.0
  description: >
    V2 目标后端契约。全部路由前缀 /api，URL 无版本号，全站无鉴权。
    工作区级 agent 模块采用 propose -> approve 语义：propose 只生成待审 artifact，
    approve 才变为当前生效 artifact；下游只读取 approved/current。
    approve 不级联重置下游，shotprompt approve 也不重建 shots。
    新的分镜链路由显式创建 shot set 触发。上游变更通过 upstreamChanged 提示表达，
    不代表 stale，也不删除候选或选择。图像/视频选择写入 image_select_artifacts /
    video_select_artifacts，每个 shot 只有一个当前 selected image/video，重复选择用 UPSERT 覆盖。
    业务语义见 docs/core/interface.md，架构见 docs/core/arc_v2.md，数据模型见 docs/core/erd.md。
servers:
  - url: http://localhost:3000
    description: 本地开发（SERVER_PORT 默认 3000）
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string }
    ShotSetId:
      name: shotSetId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: 业务错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Workspace:
      type: object
      required: [id, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        localPath: { type: string, nullable: true }
        status: { type: string }
        traceFile: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    WorkspaceInitializeResponse:
      type: object
      required: [workspace, manifest]
      properties:
        workspace:
          $ref: "#/components/schemas/Workspace"
        manifest:
          $ref: "#/components/schemas/WorkspaceManifest"
        storage:
          anyOf:
            - $ref: "#/components/schemas/StorageBinding"
            - type: "null"
        name: { type: string }
    WorkspaceDeleteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [workspaceId, deleted]
          properties:
            workspaceId: { type: string }
            deleted: { type: boolean, enum: [true] }
    WorkspaceDirectorySelectResponse:
      type: object
      required: [directory, cancelled, method]
      properties:
        directory: { type: string, nullable: true }
        cancelled: { type: boolean }
        method:
          type: string
          enum: [macos, windows, linux, unsupported]
...
paths:
...
  /api/materials:
    post:
      tags: [Material]
      summary: 登记外部商品图
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageUrl]
              properties:
                imageUrl: { type: string }
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/workspaces:
    get:
      tags: [Workspace]
      summary: 列出可见工作区（DB 已登记 + 磁盘发现的未登记草稿）
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    description: >-
                      DB 已登记且有有效本地路径的工作区（含 active storage binding）。
                      配置 WORKSPACE_DISCOVERY_ROOTS 时，仅展示这些 roots 下的 workspace。
                    items:
                      $ref: "#/components/schemas/Workspace"
                  discovered:
                    type: array
                    description: >-
                      WORKSPACE_DISCOVERY_ROOTS 下扫描到的、磁盘存在 .daireel/workspace.json
                      但 DB 无对应行的工作区（已按路径去重剔除已登记项）。点击经 init 重新打开会复用其原始 workspaceId。
                    items:
                      type: object
                      properties:
                        localPath: { type: string }
                        workspaceId:
                          type: string
                          nullable: true
                      required: [localPath]
    post:
      tags: [Workspace]
      summary: 创建逻辑工作区
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceInitializeResponse"
  /api/workspaces/directory/select:
    post:
      tags: [Workspace]
      summary: 打开本机目录选择器
      description: >-
        由桌面/本地开发环境使用。返回用户选择的本地 workspace 目录；
        不创建 workspace，不写 storage binding。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: Selected directory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDirectorySelectResponse"
  /api/workspaces/init:
    post:
      tags: [Workspace]
      summary: 按本地目录初始化工作区
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directory]
              properties:
                directory: { type: string }
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceInitializeResponse"
  /api/workspaces/{workspaceId}:
    delete:
      tags: [Workspace]
      summary: 删除已登记工作区
      description: >-
        同步 MVP：仅删除 DB 已登记 workspace。LOCAL storage 会删除工作目录下的
        .daireel/ 创作状态目录并清理该 workspace 的业务记录；不会删除工作目录中的其他文件。
        .daireel/ 已不存在时仍视为成功。当前 active storage 为 S3 时返回
        S3_WORKSPACE_DELETE_NOT_IMPLEMENTED，不清理 DB；存在 PENDING/RUNNING 生成或成片任务时返回
        WORKSPACE_DELETE_BUSY。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDeleteResponse"
        "409":
          $ref: "#/components/responses/Error"
        "501":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/status:
    get:
      tags: [Workspace]
      summary: 获取工作区状态、artifact 摘要和 active shot set
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStatus"
  /api/workspaces/{workspaceId}/storage:
    get:
      tags: [Workspace]
      summary: 获取存储绑定
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Storage binding
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    anyOf:
                      - $ref: "#/components/schemas/StorageBinding"
                      - type: "null"
  /api/workspaces/{workspaceId}/storage/bind:
    post:
      tags: [Workspace]
      summary: 绑定 LOCAL 或 S3 storage
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [kind, localPath]
                  properties:
                    kind: { type: string, enum: [local] }
                    localPath: { type: string }
                - type: object
                  required: [kind, bucket, prefix]
                  properties:
                    kind: { type: string, enum: [s3] }
                    bucket: { type: string }
                    prefix: { type: string }
                    region: { type: string }
                    endpoint: { type: string }
      responses:
        "200":
          description: Storage binding
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/StorageBinding"
        "409":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/materials:
    post:
      tags: [Material]
      summary: 上传工作区素材
      description: image/* 文件超过 10MB 返回 400 IMAGE_TOO_LARGE_FOR_MODEL；非图片仍使用通用 50MB 上限。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, dataBase64]
              properties:
                filename: { type: string }
                contentType: { type: string }
                dataBase64: { type: string }
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Uploaded asset
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/prompt-requirements:
    get:
      tags: [ModuleArtifacts]
      summary: 获取 prompt requirements 状态
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Requirements state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ModuleState"
  /api/workspaces/{workspaceId}/prompt-requirements/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 提交待审 prompt requirements
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  $ref: "#/components/schemas/PromptRequirementData"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/prompt-requirements/approve:
    post:
      tags: [ModuleArtifacts]
      summary: 批准 prompt requirements 为 current
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveArtifactRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
...
  /api/workspaces/{workspaceId}/material-intake:
    get:
      tags: [ModuleArtifacts]
      summary: 获取 material-intake 模块状态
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Module state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ModuleState"
  /api/workspaces/{workspaceId}/material-intake/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 运行物料解读 agent，生成 proposed artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/material-intake/approve:
    post:
      tags: [ModuleArtifacts]
      summary: 批准 material-intake artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveArtifactRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/product-brief:
    get:
      tags: [ModuleArtifacts]
      summary: 获取 product-brief 模块状态
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Module state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ModuleState"
  /api/workspaces/{workspaceId}/product-brief/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 运行商品卖点 agent，生成 proposed artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductBriefProposeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/product-brief/approve:
    post:
      tags: [ModuleArtifacts]
      summary: 批准 product-brief artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveArtifactRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/storyboard:
    get:
      tags: [ModuleArtifacts]
      summary: 获取 storyboard 模块状态
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Module state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ModuleState"
  /api/workspaces/{workspaceId}/storyboard/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 运行故事板 agent，生成 proposed artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/storyboard/voiceover/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 按当前三镜比例重写分镜口播，生成 proposed storyboard artifact
      description: 读取当前 approved prompt requirements、material intake、product brief，以及请求中的 15 秒三镜 storyboard draft；只重写 shots[].voiceover。请求完成前前端应保留原 draft，不显示本地压缩草稿；返回后由用户显式 approve 生效。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryboardVoiceoverProposeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/storyboard/approve:
    post:
      tags: [ModuleArtifacts]
      summary: 批准 storyboard artifact
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveArtifactRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
  /api/workspaces/{workspaceId}/shotprompt:
    get:
      tags: [ModuleArtifacts]
      summary: 获取 shotprompt 模块状态
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Module state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ModuleState"
  /api/workspaces/{workspaceId}/shotprompt/propose:
    post:
      tags: [ModuleArtifacts]
      summary: 运行分镜 prompt agent，生成 proposed artifact
      description: current approved storyboard 必须是 P0 15 秒三镜脚本；输出 shots 的数量、顺序和 index 必须与 current approved storyboard.shots 完全一致，每个 shot 包含 shotImage 与 shotVideo dict。非 P0 上游返回 400 UPSTREAM_STORYBOARD_NOT_P0。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
        "400":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/shotprompt/approve:
    post:
      tags: [ModuleArtifacts]
      summary: 批准 shotprompt artifact
      description: current approved storyboard 必须是 P0 15 秒三镜脚本；只更新 approved/current 指针，不创建或重建 shots。非 P0 上游返回 400 UPSTREAM_STORYBOARD_NOT_P0。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveArtifactRequest"
      responses:
        "200":
          $ref: "#/components/responses/ModuleArtifactEnvelope"
        "400":
          $ref: "#/components/responses/Error"
...
  /api/workspaces/{workspaceId}/shot-sets/{shotSetId}/shots:
    get:
      tags: [ShotSet, Shot]
      summary: 获取指定 shot set 下的 shots
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotSetId"
      responses:
        "200":
          description: Shots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Shot"
...